Move the cru and histalp funcs from climate to oggm.shop
