+++ docs/_code/prepare_climate.py
@@ -30,8 +30,6 @@ tasks.compute_downstream_line(gdir)
 tasks.catchment_area(gdir)
 tasks.catchment_width_geom(gdir)
 tasks.catchment_width_correction(gdir)
-data_dir = get_demo_file('HISTALP_precipitation_all_abs_1801-2014.nc')
-cfg.PATHS['cru_dir'] = os.path.dirname(data_dir)
 cfg.PARAMS['baseline_climate'] = 'HISTALP'
 cfg.PARAMS['baseline_y0'] = 1850
 tasks.process_histalp_data(gdir)
