+++ oggm/params.cfg
@@ -11,8 +11,8 @@ working_dir =
 # the empty default is what most users should do
 dem_file =
 
-# Users can specify their own climate dataset if they want to. In this case,
-# the static cru_dir will be ignored. This is useful for testing, or if you
+# Users can specify their own climate dataset if they want to. I
+# This is useful for testing, or if you
 # are simulating a single region with better data.
 # The format of the file is not (yet) very flexible. See the HISTALP data
 # in the sample-data folder for an example:
