+++ oggm/tests/test_graphics.py
@@ -12,7 +12,7 @@ gpd = pytest.importorskip('geopandas')
 # Local imports
 import oggm.utils
 from oggm.tests import mpl_image_compare
-from oggm.tests.funcs import init_hef, get_test_dir, patch_url_retrieve_github
+from oggm.tests.funcs import init_hef, get_test_dir
 from oggm import graphics
 from oggm.core import (gis, inversion, climate, centerlines, flowline,
                        massbalance)
@@ -27,17 +27,13 @@ warnings.filterwarnings("ignore", category=UserWarning,
 
 # Globals
 pytestmark = pytest.mark.test_env("graphics")
-_url_retrieve = None
 
 
-def setup_module(module):
-    module._url_retrieve = utils.oggm_urlretrieve
-    oggm.utils._downloads.oggm_urlretrieve = patch_url_retrieve_github
+def setup_module():
     graphics.set_oggm_cmaps(use_hcl=False)
 
 
-def teardown_module(module):
-    oggm.utils._downloads.oggm_urlretrieve = module._url_retrieve
+def teardown_module():
     graphics.set_oggm_cmaps()
 
 # ----------------------------------------------------------
